+++ platform-operator/experimental/module-integration/component-handler/delete/delete_handler.go
