+++ platform-operator/experimental/module-integration/component-handler/factory/factory.go
@@ -5,9 +5,9 @@ package factory
 
 import (
 	"github.com/verrazzano/verrazzano-modules/pkg/controller/handlerspi"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/module-integration/component-handler/delete"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/module-integration/component-handler/install"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/module-integration/component-handler/upgrade"
+	"github.com/verrazzano/verrazzano/platform-operator/experimental/module-integration/component-handler/delete"
+	"github.com/verrazzano/verrazzano/platform-operator/experimental/module-integration/component-handler/install"
+	"github.com/verrazzano/verrazzano/platform-operator/experimental/module-integration/component-handler/upgrade"
 )
 
 // NewModuleHandlerInfo creates a new ModuleHandlerInfo
@@ -15,6 +15,8 @@ func NewModuleHandlerInfo() handlerspi.ModuleHandlerInfo {
 	return handlerspi.ModuleHandlerInfo{
 		InstallActionHandler: install.NewHandler(),
 		DeleteActionHandler:  delete.NewHandler(),
+
+		// Update uses same handler as install
 		UpdateActionHandler:  install.NewHandler(),
 		UpgradeActionHandler: upgrade.NewHandler(),
 	}
